fix: Reduce the size of testinfra_bdd/__init__.py.
